Create users.txt
This is the "database' file which contains three users. I'm adding it so  that you don't have to register new users. First part of a line is the username, second is the hashed value of the passwords of the user (in this case all the passwords are set to "Pass@") and the third part is the balance.